refactor obs_build
